refactor(haqq-website): move dynamic link redirect to separate lazy component
